requirements.txt
torch>=2.0.0
soundfile>=0.12.1
openai-whisper>=20231117
transformers>=4.35.0
safetensors>=0.4.0
huggingface_hub>=0.17.0
irodori-tts
pydantic>=2.0.0
streamlit>=1.28.0
pydub>=0.25.1
librosa>=0.10.0
